+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+### Fixed
+- TableOutput show minimal saved time as 0
+
 ## 1.0.0 - 2016-06-06
 - First tagged version